Create ColorFinder and JiconColorFinder
The ColorFinder abstract class has been created, it has the
responsibility of retrieving the main color from an input image.
The color finding algorithm has been implemented in an extensible way,
ColorFinder is an abstract class which lets implementers deal with the
actual weighing function. JiconColorFinder is an implementation of
ColorFinder specific for this library.
The algorithm used is a porting of
https://github.com/pieroxy/color-finder.