Deal with icons containing multiple images
* JiconIcon objects structure has been modified in order to contain
a List of JiconIconImage(s), each JiconIconImage represents an image
embedded into a file. This is needed if we want to easily store all the
icon information when dealing with formats which allow to store multiple
images in the same file (see for example .ico and .gif).
* Image4j dependency removed an replaced with
com.twelvemonkeys.imageio:imageio-ico, the great advantage of this
library over Image4j is that it allows to manipulate ico files using
ImageIO standard interface.
* ImageUtil class has been created, the exposed interface allows to
easily manipulate icons containing multiple images and execute
operations for each embedded image.