Create JiconIconFactory class
JiconIcon and JiconImage constructors visibility has been changed from
public to package-private, this way library users will not be able to
build broken JiconIcon objects. Is in fact important that the order of
JiconIconImage(s) in a JiconIcon is the same of the images within the
original image file. Since the library makes this assumption the class
JiconIconFactory has been created, it exposes a method to build a
JiconIcon from an URL. Users of the library who wants to directly create
an instance of JiconIcon should use this factory.