Add .gitattributes file
The .gitattributes file has been created and added to the repository.
It marks the src/test/resources/ directory as a vendor directory.
With this change github should show Java as the main language of the
repository. Previously all the HTML files used as tests resources
were inferred as code by the github language plugin.