Major refactoring of Jicon public interface
In order to make the Jicon public API more testable, and not
tightly coupled with Jsoup, it is now possible to pass a
favicons fetching strategy to the API. A favicons fetching strategy is
simply a provider of an IconsFetcher. Jicon is not anymore a static
class and optionally takes as input the mentioned strategy.
The default strategy used when initializing Jicon with the public
constructor establishes an HTTP connection to the URL, retrieve the
Document and parses it using Jsoup.